+++ src/main/java/com/bae/rest/TeamPlayerController.java
@@ -1,0 +1,44 @@
+package com.bae.rest;
+
+import java.util.List;
+
+import com.bae.business.service.TeamPlayerService;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import com.bae.persistence.domain.TeamPlayer;
+
+
+@RestController
+public class TeamPlayerController {
+
+    private TeamPlayerService playerService;
+    private TeamPlayerController(TeamPlayerService playerService) {
+        this.playerService = playerService;
+    }
+
+    @GetMapping("/player")
+    public List<TeamPlayer> getAllPlayer() {
+        return playerService.getAllPlayer();
+    }
+
+    @PostMapping("/player")
+    public TeamPlayer addNewTrainer(@RequestBody TeamPlayer player) {
+        return playerService.addNewPlayer(player);
+    }
+
+    @PutMapping("/player")
+    public TeamPlayer updateTrainer(@RequestBody TeamPlayer player) {
+        return playerService.updatePlayer(player);
+    }
+
+    @DeleteMapping("/player/{id}")
+    public String deletePlayer(@PathVariable long id) {
+        return "Player deleted";
+    }
+}
